feat(paywall): render the billing embed inline in the card, not a dialog
The card that replaces the composer is now the iframe itself: header row
(title + Dismiss) with the platform billing page below, tall enough to
show the top-up panel. Loads on mount for admins on web/cloud; Electron,
members and non-cloud web keep the external CTA.
