+++ src/renderer/components/provider-error/billing-embed-frame.tsx
@@ -1,0 +1,108 @@
+import { useEffect, useRef, useState } from 'react'
+import { Loader2 } from 'lucide-react'
+
+import { Button } from '@renderer/components/ui/button'
+import { useBillingEmbedSession } from '@renderer/hooks/use-billing-embed'
+import { captureRendererException } from '@renderer/lib/error-reporting'
+import { openExternalUrl } from '@renderer/lib/open-external'
+
+// Must match the platform's /embed contract (apps/web/src/lib/billing-embed.ts).
+export const BILLING_EMBED_MESSAGE_TYPE = 'gamut-billing-embed'
+type BillingEmbedEvent = 'ready' | 'billing-updated' | 'session-expired'
+
+function readEmbedEvent(data: unknown): BillingEmbedEvent | null {
+  if (typeof data !== 'object' || data === null) return null
+  const record = data as { type?: unknown; event?: unknown }
+  if (record.type !== BILLING_EMBED_MESSAGE_TYPE) return null
+  return record.event === 'ready' || record.event === 'billing-updated' || record.event === 'session-expired'
+    ? record.event
+    : null
+}
+
+export interface BillingEmbedFrameProps {
+  intent?: 'topup'
+  /** External billing URL for the fallback button. */
+  fallbackHref: string | null
+  /** Fired on every `billing-updated` from the platform page. */
+  onBillingUpdated: () => void
+}
+
+// Web paywall body: the platform billing page inline, where the composer was.
+// The iframe URL is a one-time session bootstrap minted by the host on mount.
+// Anything that stops the embed from working falls back to opening billing
+// externally, exactly like the Electron flow.
+export function BillingEmbedFrame({ intent, fallbackHref, onBillingUpdated }: BillingEmbedFrameProps) {
+  const session = useBillingEmbedSession()
+  const [frameReady, setFrameReady] = useState(false)
+  const [expired, setExpired] = useState(false)
+  const requested = useRef(false)
+  const { mutate } = session
+
+  useEffect(() => {
+    if (requested.current) return
+    requested.current = true
+    mutate({ intent })
+  }, [intent, mutate])
+
+  const platformOrigin = session.data?.platformOrigin ?? null
+  useEffect(() => {
+    if (!platformOrigin) return
+    const onMessage = (event: MessageEvent) => {
+      if (event.origin !== platformOrigin) return
+      const embedEvent = readEmbedEvent(event.data)
+      if (embedEvent === 'ready') setFrameReady(true)
+      else if (embedEvent === 'billing-updated') onBillingUpdated()
+      else if (embedEvent === 'session-expired') setExpired(true)
+    }
+    window.addEventListener('message', onMessage)
+    return () => window.removeEventListener('message', onMessage)
+  }, [platformOrigin, onBillingUpdated])
+
+  const sessionError = session.error
+  useEffect(() => {
+    if (!sessionError) return
+    console.warn('[Paywall] billing embed unavailable:', sessionError)
+    captureRendererException(sessionError, { tags: { area: 'paywall', op: 'billing-embed', code: sessionError.code } })
+  }, [sessionError])
+
+  const failure = expired ? 'This billing session has expired.' : sessionError?.message ?? null
+
+  return (
+    <div
+      className="relative h-[min(560px,62vh)] w-full overflow-hidden rounded-lg border bg-background"
+      data-testid="billing-embed-body"
+    >
+      {failure ? (
+        <div className="flex h-full flex-col items-center justify-center gap-3 px-6 text-center">
+          <p className="text-sm text-muted-foreground">{failure}</p>
+          {fallbackHref && (
+            <Button size="sm" onClick={() => void openExternalUrl(fallbackHref)}>
+              Open billing in a new tab
+            </Button>
+          )}
+        </div>
+      ) : (
+        <>
+          {session.data && (
+            <iframe
+              title="Workspace billing"
+              src={session.data.embedUrl}
+              className="h-full w-full border-0"
+              referrerPolicy="strict-origin"
+              data-testid="billing-embed-frame"
+            />
+          )}
+          {!frameReady && (
+            <div
+              className="absolute inset-0 flex items-center justify-center gap-2 bg-background text-xs text-muted-foreground"
+              data-testid="billing-embed-loading"
+            >
+              <Loader2 className="h-3.5 w-3.5 animate-spin" aria-hidden="true" />
+              Loading billing…
+            </div>
+          )}
+        </>
+      )}
+    </div>
+  )
+}
